fixing type's func places
